% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.R
\name{sdt_loss}
\alias{sdt_loss}
\title{Computing the MSE}
\usage{
sdt_loss(
  pars,
  data,
  method = c("expectation", "simulation"),
  ntrials = 1000,
  log_mse = TRUE
)
}
\arguments{
\item{pars}{Numeric, should be a list of initial values for the response bias
and the internal noise.}

\item{data}{Dataframe, with observed prop_agree and prop_first.}

\item{method}{Character, the method for computing prop_agree and prop_first ("simulation" or "expectation").}

\item{ntrials}{Numeric, number of trials per block (defaults to 1e4).}

\item{log_mse}{Logical, should we return the log-MSE (instead of the MSE).}
}
\description{
Computing the MSE from prop_agree and prop_first.
}
\examples{
\dontrun{
# generating prop_agree and prop_first from pars
sdt_df <- sdt_data(pars = c(0, 1), ntrials = 1e4)

# computing the MSE loss (or log-MSE)
sdt_loss(par = c(0, 1), data = sdt_df, ntrials = 1e4)
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@cnrs.fr}.
}
