% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_produced_speech}
\alias{self_produced_speech}
\title{Self-produced speech revcorr data}
\format{
A data frame with 1400 rows and 7 variables:
\describe{
\item{participant}{...}
\item{sex}{...}
\item{age}{...}
\item{block}{...}
\item{trial}{...}
\item{response}{...}
\item{RT}{...}
}
}
\source{
Home-made revcorr speech data.
}
\usage{
self_produced_speech
}
\description{
Self-produced speech revcorr data.
}
\keyword{datasets}
