% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_produced_speech_full}
\alias{self_produced_speech_full}
\title{Full version of the self-produced speech revcorr data}
\format{
A data frame with 72800 rows and 7 variables:
\describe{
\item{participant}{...}
\item{block}{...}
\item{trial}{...}
\item{resp}{...}
\item{RT}{...}
\item{f}{...}
\item{eq}{...}
}
}
\source{
Home-made revcorr speech data.
}
\usage{
self_produced_speech_full
}
\description{
Full version of the self-produced speech revcorr data, with properties of
the filters at each trial.
}
\keyword{datasets}
