% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_voice}
\alias{self_voice}
\title{Full version of the self-voice data}
\format{
A data frame with 1931916 rows and 7 variables:
\describe{
\item{participant}{...}
\item{block}{...}
\item{trial}{...}
\item{resp}{...}
\item{RT}{...}
\item{f}{...}
\item{eq}{...}
}
}
\source{
Reverse correlation speech data collected by Joao Sarzedas et al. (in preparation).
}
\usage{
self_voice
}
\description{
Full version of the Sarzedas et al. (in preparation) speech revcorr data,
with properties of the filters at each trial.
}
\keyword{datasets}
