% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{computing_kernel}
\alias{computing_kernel}
\title{Computing the kernel from 2AFC data}
\usage{
computing_kernel(
  data,
  method = c("difference", "glm", "glmm"),
  double_pass = TRUE
)
}
\arguments{
\item{data}{Dataframe, containing the data...}

\item{method}{Character, the computation method, either "difference", "glm", or "glmm".}

\item{double_pass}{Logical, indicating whether the last block was repeated.}
}
\value{
The original data augmented with the kernel.
}
\description{
Computing the kernel from 2AFC data. It should be noted that when using the
"glm" method, the group-level (average) kernel is simply averaged from the
individual-level kernels. To obtain \strong{shrinked} (i.e., adjusted) group-level
kernels, one needs to specify the "glmm" method (work in progress).
}
\examples{
\dontrun{
# importing the data
data(self_produced_speech_full)
head(self_produced_speech_full)

# computing the kernel
computing_kernel(data = self_produced_speech_full, method = "difference")

# plotting it
computing_kernel(data = self_produced_speech_full, method = "difference") |> plot()
computing_kernel(data = self_produced_speech_full, method = "glm") |> plot()
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
