% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddm_fitting}
\alias{ddm_fitting}
\title{Fitting the full DDM}
\usage{
ddm_fitting(rt, resp, method = c("nlminb", "optim"))
}
\arguments{
\item{rt}{Numeric, the response time (in \strong{seconds}).}

\item{resp}{Numeric, a vector indicating the chosen stimulus at each trial.}

\item{method}{Character, the optimisation method (i.e., "nlminb" or "optim").}
}
\value{
The optimised parameter values and further output from nlminb() or optim().
}
\description{
Fitting the full DDM using the \code{fddm} package.
}
\examples{
\dontrun{
# importing the data
data(self_produced_speech)
head(self_produced_speech)

# reshaping the data
df <- self_produced_speech |>
  # removing the last block
  filter(block < 7) |>
  # keeping only the relevant columns
  select(participant, choice = response, RT) |>
  mutate(choice = ifelse(test = choice == "stim1", yes = 1, no = 2) )

# splitting data by participant
df_ln <- df |> filter(participant == "LN")
df_mm <- df |> filter(participant == "MM")

# fitting the full DDM for LN and MM (pars are a, v, t0, w, sv)
ddm_fitting(rt = df_ln$RT, resp = df_ln$choice, method = "nlminb")$par
ddm_fitting(rt = df_mm$RT, resp = df_mm$choice, method = "nlminb")$par
}
}
