% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddm_log_likelihood}
\alias{ddm_log_likelihood}
\title{Log-likelihood of the DDM}
\usage{
ddm_log_likelihood(rt, resp, par)
}
\arguments{
\item{rt}{Numeric, the response time (in \strong{seconds}).}

\item{resp}{Numeric, a vector indicating the chosen stimulus at each trial.}

\item{par}{Character, the optimisation method (i.e., "nlminb" or "optim").}
}
\value{
The negative log-likelihood.
}
\description{
Computing the (negative) log-likelihood of the DDM using the \code{fddm} package.
}
\examples{
\dontrun{
# importing the data
data(self_produced_speech)
head(self_produced_speech)

# reshaping the data
df <- self_produced_speech |>
  # removing the last block
  filter(block < 7) |>
  # keeping only the relevant columns
  select(participant, choice = response, RT) |>
  mutate(choice = ifelse(test = choice == "stim1", yes = 1, no = 2) )

# splitting data by participant
df_ln <- df |> filter(participant == "LN")
df_mm <- df |> filter(participant == "MM")

# computing the (negative) log-likelihood given some data and parameter values
# pars are a, v, t0, w, sv
ddm_log_likelihood(rt = df_ln$RT, resp = df_ln$choice, par = c(1, 1, 0, 0.5, 0) )
ddm_log_likelihood(rt = df_mm$RT, resp = df_mm$choice, par = c(1, 1, 0, 0.5, 0) )
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
