% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddm_log_likelihood}
\alias{ddm_log_likelihood}
\title{Log-likelihood of the DDM}
\usage{
ddm_log_likelihood(pars, rt, resp)
}
\arguments{
\item{pars}{Numeric, should be a list of initial values for the parameters.}

\item{rt}{Numeric, the response time (in \strong{seconds}).}

\item{resp}{Numeric, a vector indicating the chosen stimulus at each trial.}
}
\value{
The negative log-likelihood.
}
\description{
Computing the (negative) log-likelihood of the DDM using the \code{fddm} package.
}
\examples{
\dontrun{
# importing the self-voice data
data(self_voice)
head(self_voice)

# reshaping the data
df <- self_voice |>
  # removing the last block
  filter(block < max(block) ) |>
  # keeping only the relevant columns
  select(participant, choice = resp, RT) |>
  mutate(choice = ifelse(test = choice == "stim1", yes = 1, no = 2) )

# splitting data by participant (keeping only the first one)
df_ppt <- df |> filter(participant == unique(participant)[1])

# computing the (negative) log-likelihood given some data and parameter values
# pars are a, v, t0, w, sv
ddm_log_likelihood(rt = df_ppt$RT, resp = df_ppt$choice, par = c(1, 1, 0, 0.5, 0) )
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
