% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{response_consistency}
\alias{response_consistency}
\title{Computing different metrics of response's consistency.}
\usage{
response_consistency(data, double_pass = TRUE)
}
\arguments{
\item{data}{Dataframe, with reverse correlation data. Should contain the
following columns: participant, block, trial, resp, f, eq.}

\item{double_pass}{Logical, indicating whether the last block was repeated.}
}
\value{
Numeric, the MSE.
}
\description{
Computing different metrics of response's consistency from both double-pass
and non double-pass reverse correlation data.
}
\examples{
\dontrun{
# importing the self-voice data
data(self_voice)
head(self_voice)

# computing metrics of response consistency per participant and per block
response_consistency(self_voice) |> head(10)
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
