% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.R
\name{sdt_data}
\alias{sdt_data}
\title{Generating data from the SDT model}
\usage{
sdt_data(
  pars,
  return_summary = TRUE,
  method = c("simulation", "expectation"),
  ntrials = 10000
)
}
\arguments{
\item{pars}{Numeric, should be a list of initial values for the response bias
and the internal noise.}

\item{return_summary}{Logical, should we return only prop_agree and prop_first, or the full data.}

\item{method}{Character, the method for computing prop_agree and prop_first ("simulation" or "expectation").}

\item{ntrials}{Numeric, number of trials per block (defaults to 1e4).}
}
\value{
Numeric, the MSE.
}
\description{
Generating data from the SDT model adapted from
Goupil et al. (2021), \url{https://doi.org/10.1038/s41467-020-20649-4}.
}
\examples{
\dontrun{
# generating prop_agree and prop_first from parameters values
sdt_data(pars = c(0, 1), ntrials = 1e4, method = "simulation")

# generating prop_agree and prop_first from parameters values
sdt_data(pars = c(0, 1), ntrials = 1e4, method = "expectation")
}

}
\references{
Goupil, L., Ponsot, E., Richardson, D. et al. (2021). Listeners'
perceptions of the certainty and honesty of a speaker are associated with a
common prosodic signature. \emph{Nature Communications 12}, 861. \url{https://doi.org/10.1038/s41467-020-20649-4}.
}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
