% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.R
\name{sdt_data}
\alias{sdt_data}
\title{Generating data from the SDT model}
\usage{
sdt_data(pars, ntrials = 10000)
}
\arguments{
\item{pars}{Numeric, should be a list of initial values for the response bias
and the internal noise.}

\item{ntrials}{Numeric, number of trials per block (defaults to 1e4).}
}
\value{
Numeric, the MSE.
}
\description{
Generating data from the SDT model and computing the MSE. Adapted from
Goupil et al. (2021), \url{https://doi.org/10.1038/s41467-020-20649-4}.
}
\examples{
\dontrun{
# example data (70\% agreement and no response bias)
sdt_df <- data.frame(prop_agree = 0.7, prop_first = 0.5, ntrials = 1e2)

# generating prop_agree and prop_first
sdt_data(par = c(0, 1), data = sdt_df)
}

}
\references{
Goupil, L., Ponsot, E., Richardson, D. et al. (2021). Listeners'
perceptions of the certainty and honesty of a speaker are associated with a
common prosodic signature. \emph{Nature Communications 12}, 861. \url{https://doi.org/10.1038/s41467-020-20649-4}.
}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
