% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.R
\name{sdt_fitting}
\alias{sdt_fitting}
\title{Fitting SDT model}
\usage{
sdt_fitting(
  data,
  ntrials = 10000,
  method = c("nlminb", "SANN", "Nelder-Mead", "CG", "BFGS", "bobyqa", "grid"),
  grid_res = 0.01
)
}
\arguments{
\item{data}{Dataframe, with observed prop_agree and prop_int1.}

\item{ntrials}{Numeric, number of trials in the SDT model.}

\item{method}{Character, the optimisation method, see possible values below.
Beware that method "grid" can take some time, depending on the size of the grid.}

\item{grid_res}{Numeric, grid resolution in units of response bias or internal
noise (only used for method "grid").}
}
\value{
The optimised parameter values and further convergence information.
}
\description{
Fitting the SDT model. NB: Best results are obtained with the "bobyqa"
method and at least 1e5 trials.
}
\details{
Add some doc here...
}
\examples{
\dontrun{
# actual observations from pilot trials (LN and MM)
df_ln <- data.frame(prop_agree = 0.7, prop_int1 = 0.413, ntrials = 1e2)
df_mm <- data.frame(prop_agree = 0.58, prop_int1 = 0.563, ntrials = 1e2)

# fitting the SDT model for LN and MM
sdt_fitting(data = df_ln, ntrials = 1e5, method = "bobyqa")
sdt_fitting(data = df_mm, ntrials = 1e5, method = "bobyqa")

# for info: results obtained via the method of Ponsot (not in the package yet)
# for LN
# The minimised response bias value is: -0.227
# The minimised internal noise value is: 0.833
# for MM
# The minimised response bias value is: 0.227
# The minimised internal noise value is: 1.566
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
