% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt.R
\name{sdt_loss}
\alias{sdt_loss}
\title{Computing the MSE}
\usage{
sdt_loss(pars, data, ntrials = 10000, log_mse = TRUE)
}
\arguments{
\item{pars}{Numeric, should be a list of initial values for the response bias
and the internal noise.}

\item{data}{Dataframe, with observed prop_agree and prop_first.}

\item{ntrials}{Numeric, number of trials per block (defaults to 1e4).}

\item{log_mse}{Logcial, should we return the log-MSE (instead of the MSE).}
}
\description{
Computing the MSE from prop_agree and prop_first.
}
\examples{
\dontrun{
# example data (70\% agreement and no response bias)
sdt_df <- data.frame(prop_agree = 0.7, prop_first = 0.5, ntrials = 1e2)

# generating prop_agree and prop_first
sdt_data(par = c(0, 1), data = sdt_df)
}

}
\author{
Ladislas Nalborczyk \email{ladislas.nalborczyk@gmail.com}.
}
