% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_voice}
\alias{self_voice}
\title{Subset of the self-voice data}
\format{
A data frame with 104979 rows and 7 variables:
\describe{
\item{participant}{Character variable indicating the participant's ID.}
\item{block}{Integer variable indicating the block number.}
\item{trial}{Integer variable indicating the trial number.}
\item{response}{Integer variable indicating the participant's response.}
\item{RT}{Numeric variable indicating the reaction time (in seconds).}
\item{feature}{Numeric variable indicating the feature ID.}
\item{value}{Numeric variable indicating the value of the feature.}
}
}
\source{
Reverse correlation speech data collected by Joao Sarzedas et al. (in preparation).
}
\usage{
self_voice
}
\description{
Subset of the speech reverse correlation data from Sarzedas et al. (in preparation).
}
\keyword{datasets}
