% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_voice}
\alias{self_voice}
\title{Subset of the self-voice data}
\format{
A data frame with 104979 rows and 7 variables:
\describe{
\item{participant}{...}
\item{block}{...}
\item{trial}{...}
\item{response}{...}
\item{RT}{...}
\item{feature}{...}
\item{value}{...}
}
}
\source{
Reverse correlation speech data collected by Joao Sarzedas et al. (in preparation).
}
\usage{
self_voice
}
\description{
Subset of the spech revcorr data from Sarzedas et al. (in preparation),
with properties of the filters at each trial.
}
\keyword{datasets}
